+++ core/includes/page_queue.h
@@ -1,8 +1,4 @@
-#ifndef PAGE_QUEUE_H
-#define PAGE_QUEUE_H
-#endif
-
-#include "cs_page.h"
+#include "pageProcesser.h"
 
 void page_queue_init(void);
