+++ core/pageProcesser/cs_page.c
@@ -1,5 +1,7 @@
 #include "cs_page.h"
 
+
+
 /*
   clear_page: fill the data with all zeros.
 */
